requirements.txt
black==25.1.0
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
et_xmlfile==2.0.0
mypy_extensions==1.1.0
numpy==2.2.5
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
platformdirs==4.3.7
pycparser==2.22
PyMuPDF==1.25.5
pypdf==5.4.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
